refactor: pull up icons from `core` to `media` pack
