--- packages/core/src/icon/EyeClosedIcon.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const EyeClosedIcon: React.SFC = () => (
-  <svg viewBox="0 0 24 24" focusable="false">
-    <path
-      d="M17.7053697,16.2911561 L19.7071068,18.2928932 C20.0976311,18.6834175 20.0976311,19.3165825 19.7071068,19.7071068 C19.3165825,20.0976311 18.6834175,20.0976311 18.2928932,19.7071068 L4.29289322,5.70710678 C3.90236893,5.31658249 3.90236893,4.68341751 4.29289322,4.29289322 C4.68341751,3.90236893 5.31658249,3.90236893 5.70710678,4.29289322 L8.17802973,6.76381617 C9.31365126,6.29913303 10.5916396,6 12,6 C18.4116667,6 22,12 22,12 C22,12 20.5079811,14.5175853 17.7053697,16.2911561 Z M9.74237655,8.32816299 L10.7065845,9.29237099 C11.0982198,9.10495216 11.5368535,9 12,9 C13.6568542,9 15,10.3431458 15,12 C15,12.4631465 14.8950478,12.9017802 14.707629,13.2934155 L16.2508489,14.8366354 C16.8914793,14.4691308 17.5049066,14.0212 18.0909003,13.5024621 C18.6307336,13.0245867 19.1124581,12.5157305 19.5314344,12.0036733 C19.108912,11.4896187 18.6236353,10.9787665 18.0804933,10.4991122 C16.2953932,8.92267152 14.2646189,8 12,8 C11.2233763,8 10.4704171,8.11321363 9.74237655,8.32816299 Z M13.4034596,17.9034596 C12.9521366,17.9661228 12.4842224,18 12,18 C6.01666667,18 2,12 2,12 C2,12 2.89740707,10.5417895 4.56312283,9.06312283 L5.98022077,10.4802208 C5.9711024,10.4881689 5.96198951,10.4961337 5.95288229,10.5041152 C5.42140341,10.9698991 4.94246612,11.4650567 4.52111756,11.9642277 C4.97808057,12.4828798 5.49252447,12.9975125 6.05693989,13.4807219 C7.7601339,14.9388666 9.58988601,15.8400891 11.480671,15.980671 L13.4034596,17.9034596 Z"
-    />
-  </svg>
-)
-
-export default EyeClosedIcon
