--- packages/core/src/icon/EyeOpenedIcon.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const EyeOpenedIcon: React.SFC = () => (
-  <svg viewBox="0 0 24 24" focusable="false">
-    <path
-      d="M22,12 C22,12 18.4441667,18 12,18 C6.01666667,18 2,12 2,12 C2,12 5.6925,6 12,6 C18.4116667,6 22,12 22,12 Z M18.0804933,10.4991122 C16.2953932,8.92267152 14.2646189,8 12,8 C9.78142959,8 7.7559762,8.92389777 5.95288229,10.5041152 C5.42140341,10.9698991 4.94246612,11.4650567 4.52111756,11.9642277 C4.97808057,12.4828798 5.49252447,12.9975125 6.05693989,13.4807219 C7.91562552,15.0719867 9.92503211,16 12,16 C14.2789762,16 16.311426,15.0777022 18.0909003,13.5024621 C18.6307336,13.0245867 19.1124581,12.5157305 19.5314344,12.0036733 C19.108912,11.4896187 18.6236353,10.9787665 18.0804933,10.4991122 Z M12,15 C10.3431458,15 9,13.6568542 9,12 C9,10.3431458 10.3431458,9 12,9 C13.6568542,9 15,10.3431458 15,12 C15,13.6568542 13.6568542,15 12,15 Z M12,13 C12.5522847,13 13,12.5522847 13,12 C13,11.4477153 12.5522847,11 12,11 C11.4477153,11 11,11.4477153 11,12 C11,12.5522847 11.4477153,13 12,13 Z"
-    />
-  </svg>
-)
-
-export default EyeOpenedIcon
