--- packages/core/src/icon/QuestionIcon.tsx
@@ -1,18 +1,0 @@
-import React from 'react'
-
-const QuestionIcon: React.SFC = () => (
-  <svg viewBox="0 0 24 24" focusable="false">
-    <path
-      fill="#FFF"
-      stroke="#000"
-      strokeOpacity="0.15"
-      d="M21.5 12c0-5.245-4.255-9.5-9.5-9.5S2.5 6.755 2.5 12s4.255 9.5 9.5 9.5 9.5-4.255 9.5-9.5z"
-    />
-    <path
-      fill="#000"
-      d="M10.005 8.059c.053-.256.12-.326.305-.447.186-.121.385-.228.598-.32.213-.091.442-.163.687-.215.245-.051.51-.077.796-.077.401 0 .762.056 1.084.168.322.112.596.272.822.481.226.208.4.46.52.753.122.294.183.62.183.979 0 .34-.045.635-.134.884a2.503 2.503 0 0 1-.79 1.126c-.164.135-.32.26-.467.374a3.681 3.681 0 0 0-.383.34.767.767 0 0 0-.208.376l-.049.257c-.072.378-.444.684-.84.684h.064c-.392 0-.751-.319-.8-.692l-.05-.39a.978.978 0 0 1 .064-.575c.076-.164.182-.312.316-.444s.285-.258.451-.377c.166-.119.322-.246.467-.383.145-.137.267-.29.367-.457.1-.169.15-.37.15-.602 0-.27-.084-.483-.252-.643-.168-.159-.4-.238-.694-.238a1.85 1.85 0 0 0-.956.238c-.108.061-.203.116-.284.165a.464.464 0 0 1-.243.074.437.437 0 0 1-.41-.249s-.367-.534-.314-.79zM11 16a.995.995 0 0 1 .602-.921A.988.988 0 0 1 12 15c.14 0 .27.026.392.079.121.053.227.124.318.213A.969.969 0 0 1 13 16c0 .14-.026.272-.077.395a.961.961 0 0 1-.531.529c-.121.05-.252.076-.392.076a1.02 1.02 0 0 1-.398-.076.962.962 0 0 1-.525-.53A1.016 1.016 0 0 1 11 16z"
-    />
-  </svg>
-)
-
-export default QuestionIcon
