--- packages/core/src/icon/RepeatIcon.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-const RepeatIcon: React.SFC = () => (
-  <svg viewBox="0 0 24 24" focusable="false">
-    <path
-      d="M19.065 5.924a1 1 0 0 1-1.57 1.24A7 7 0 0 0 5 11.5a1 1 0 1 1-1.999 0 9 9 0 0 1 16.065-5.576zM5.01 18.17a1 1 0 0 1 1.553-1.26A7 7 0 0 0 19 12.5a1 1 0 0 1 2 0 9 9 0 0 1-15.99 5.67z"
-    />
-    <path
-      d="M20 11.914l-1.293 1.293a1 1 0 0 1-1.414-1.414l2-2a1 1 0 0 1 1.414 0l2 2a1 1 0 1 1-1.414 1.414L20 11.914zM4 12.086l1.293-1.293a1 1 0 0 1 1.414 1.414l-2 2a1 1 0 0 1-1.414 0l-2-2a1 1 0 1 1 1.414-1.414L4 12.086z"
-    />
-  </svg>
-)
-
-export default RepeatIcon
