--- packages/core/src/icon/StarIcon.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const StarIcon: React.SFC = () => (
-  <svg viewBox="0 0 24 24" focusable="false">
-    <path
-      d="M16.108 20.347l-3.983-2.064a.272.272 0 0 0-.25 0l-3.983 2.064a1.35 1.35 0 0 1-1.413-.101 1.306 1.306 0 0 1-.534-1.294l.76-4.37a.262.262 0 0 0-.077-.234l-3.222-3.096a1.301 1.301 0 0 1-.34-1.356 1.328 1.328 0 0 1 1.084-.9l4.453-.637a.268.268 0 0 0 .202-.145l1.992-3.977A1.336 1.336 0 0 1 12 3.5c.515 0 .976.283 1.204.737l1.99 3.977c.04.078.116.132.203.144l4.453.638c.51.073.924.418 1.083.9a1.301 1.301 0 0 1-.34 1.356l-3.221 3.095a.262.262 0 0 0-.078.235l.761 4.37c.087.5-.118.996-.534 1.294a1.35 1.35 0 0 1-1.413.1zM12 16.265c.176 0 .352.042.513.125l3.27 1.695c.11.056.2.007.232-.017a.21.21 0 0 0 .088-.212l-.624-3.59a1.076 1.076 0 0 1 .316-.96l2.646-2.542a.21.21 0 0 0 .056-.223.213.213 0 0 0-.178-.148l-3.656-.523a1.1 1.1 0 0 1-.83-.594L12.198 6.01A.215.215 0 0 0 12 5.89a.215.215 0 0 0-.198.12l-1.635 3.265a1.1 1.1 0 0 1-.83.595l-3.656.523a.214.214 0 0 0-.178.148.21.21 0 0 0 .056.223l2.646 2.541c.26.25.378.61.317.962l-.625 3.589a.21.21 0 0 0 .088.212.218.218 0 0 0 .232.017l3.27-1.695c.16-.083.337-.125.513-.125z"
-    />
-  </svg>
-)
-
-export default StarIcon
