+++ packages/core/src/icon/index.ts
@@ -1,16 +1,1 @@
-import Icon from './Icon'
-
-import EyeClosedIcon from './EyeClosedIcon'
-import EyeOpenedIcon from './EyeOpenedIcon'
-import QuestionIcon from './QuestionIcon'
-import RepeatIcon from './RepeatIcon'
-import StarIcon from './StarIcon'
-
-export {
-  Icon,
-  EyeClosedIcon,
-  EyeOpenedIcon,
-  QuestionIcon,
-  RepeatIcon,
-  StarIcon,
-}
+export {Icon} from '@qiwi/pijma-media'
