+++ packages/desktop/package.json
@@ -16,6 +16,7 @@
   ],
   "dependencies": {
     "@qiwi/pijma-core": "^1.0.0",
+    "@qiwi/pijma-media": "^1.0.0",
     "emotion": "^9.2.6",
     "emotion-theming": "^9.2.3",
     "polished": "^1.9.2",
@@ -26,7 +27,6 @@
     "tslib": "^1.9.2"
   },
   "devDependencies": {
-    "@qiwi/pijma-media": "^1.0.0",
     "@types/react": "^16.3.17",
     "@types/react-dom": "^16.0.6",
     "@types/react-text-mask": "^5.4.0",
