+++ packages/mobile/package.json
@@ -16,6 +16,7 @@
   ],
   "dependencies": {
     "@qiwi/pijma-core": "^1.0.0",
+    "@qiwi/pijma-media": "^1.0.0",
     "emotion": "^9.2.6",
     "polished": "^1.9.2",
     "react": "^16.4.1",
